Print Method for evaluCATE Objects

Prints an evaluCATE object.

# S3 method for evaluCATE
print(x, latex = "none", ...)

Arguments

x

An evaluCATE object.

latex

String to denote whether and which table to print in LATEX code. Must be one of "none", "BLP".

...

Further arguments passed to or from other methods.

Value

Prints an evaluCATE object.

Details

Compilation of the LATEX code requires the following packages: booktabs, float, adjustbox.

Author

Riccardo Di Francesco

Examples

## Generate data.
set.seed(1986)

n <- 1000
k <- 2

X <- matrix(rnorm(n * k), ncol = k)
colnames(X) <- paste0("x", seq_len(k))
D <- rbinom(n, size = 1, prob = 0.5)
mu0 <- 0.5 * X[, 1]
mu1 <- 0.5 * X[, 1] + X[, 2]
Y <- mu0 + D * (mu1 - mu0) + rnorm(n)

## Sample split.
train_idx <- sample(c(TRUE, FALSE), length(Y), replace = TRUE)

X_tr <- X[train_idx, ]
X_val <- X[!train_idx, ]

D_tr <- D[train_idx]
D_val <- D[!train_idx]

Y_tr <- Y[train_idx]
Y_val <- Y[!train_idx]

## CATEs estimation.
library(grf)

forest <- causal_forest(X_tr, Y_tr, D_tr) # We use only the training sample.
cates <- predict(forest, X)$predictions # We predict on the whole sample.

## CATEs evaluation. Estimate all nuisances internally. 
pscore <- rep(0.5, length(Y))
evaluation <- evalue_cates(Y, D, X, cates, train_idx, pscore = pscore)
#> Estimating nuisance functions and AIPW scores; 
#> BLP estimation; 
#> GATES estimation; 
#> RATE estimation; 
#> Output. 
#> 

## Print.
print(evaluation)
#> CATEs evaluation results 
#> 
#> Estimated ATE + 95% confidence intervals: 
#> wr_none : -0.06 [-0.274, 0.147] 
#> wr_cddf1: -0.11 [-0.31, 0.093] 
#> wr_cddf2: -0.07 [-0.248, 0.109] 
#> wr_mck1 : -0.09 [-0.263, 0.093] 
#> ht_none : -0.06 [-0.274, 0.147] 
#> ht_cddf1: -0.11 [-0.31, 0.093] 
#> ht_cddf2: -0.07 [-0.248, 0.109] 
#> ht_mck1 : -0.07 [-0.253, 0.106] 
#> ht_mck2 : -0.07 [-0.247, 0.109] 
#> ht_mck3 : -0.07 [-0.242, 0.112] 
#> aipw    : -0.04 [-0.231, 0.143] 
#> 
#> Estimated HET + 95% confidence intervals: 
#> wr_none : 1.03 [0.784, 1.283] 
#> wr_cddf1: 1.07 [0.826, 1.304] 
#> wr_cddf2: 1.07 [0.876, 1.259] 
#> wr_mck1 : 1.05 [0.861, 1.24] 
#> ht_none : 1.03 [0.784, 1.283] 
#> ht_cddf1: 1.07 [0.826, 1.304] 
#> ht_cddf2: 1.07 [0.876, 1.259] 
#> ht_mck1 : 1.06 [0.865, 1.249] 
#> ht_mck2 : 1.06 [0.87, 1.251] 
#> ht_mck3 : 1.06 [0.867, 1.248] 
#> aipw    : 1.06 [0.859, 1.265] 
#> 
#> RATEs results + 95% confidence intervals: 
#> AUTOC: 0.88 [0.697, 1.061]
#> QINI:  0.28 [0.221, 0.332] 
print(evaluation, latex = "BLP")
#> \begingroup
#>     \setlength{\tabcolsep}{8pt}
#>     \renewcommand{\arraystretch}{1.1}
#>     \begin{table}[H]
#>         \centering
#>         \begin{adjustbox}{width = 1\textwidth}
#>         \begin{tabular}{@{\extracolsep{5pt}}l c c c c c c c c c c c}
#>         \\[-1.8ex]\hline
#>         \hline \\[-1.8ex]
#>         & wr\_none & wr\_cddf1 & wr\_cddf2 & wr\_mck1 & ht\_none & ht\_cddf1 & ht\_cddf2 & ht\_mck1 & ht\_mck2 & ht\_mck3 & aipw \\
#>         \addlinespace[2pt]
#>         \hline \\[-1.8ex] 
#> 
#>         ATE ($\beta_1$) & -0.06 & -0.11 & -0.07 & -0.09 & -0.06 & -0.11 & -0.07 & -0.07 & -0.07 & -0.07 & -0.04 \\ 
#>                          &  [-0.274, 0.147] & [-0.31, 0.093] & [-0.248, 0.109] & [-0.263, 0.093] & [-0.274, 0.147] & [-0.31, 0.093] & [-0.248, 0.109] & [-0.253, 0.106] & [-0.247, 0.109] & [-0.242, 0.112] & [-0.231, 0.143] \\ 
#>         HET ($\beta_2$) & 1.03 & 1.07 & 1.07 & 1.05 & 1.03 & 1.07 & 1.07 & 1.06 & 1.06 & 1.06 & 1.06 \\ 
#>                          &  [0.784, 1.283] & [0.826, 1.304] & [0.876, 1.259] & [0.861, 1.24] & [0.784, 1.283] & [0.826, 1.304] & [0.876, 1.259] & [0.865, 1.249] & [0.87, 1.251] & [0.867, 1.248] & [0.859, 1.265] \\ 
#> 
#>         \addlinespace[3pt]
#>         \\[-1.8ex]\hline
#>         \hline \\[-1.8ex]
#>         \end{tabular}
#>         \end{adjustbox}
#>         \caption{BLP results. $95\%$ confidence intervals are displayed in brackets under each point estimate.}
#>         \label{table_blp_results}
#>     \end{table}
#> \endgroup