Testing for Effect Heterogeneity using the estimated GATES

Tests for effect heterogeneity using the estimated parametric model for the sorted group average treatment effects (GATES).

test_parametric_gates(model)

Arguments

model

Estimated parametric model, as one of those returned by gates_estimation.

Value

A list with the p-values for the three hypotheses tested.

Details

model must consist of a lm_robust object where the coefficients identifying the GATES must be called "group1", "group2", and so on.

Three distinct hypotheses of effect heterogeneity are tested: whether all GATES are equal to each other, whether the largest and the smallest GATES are different from each other, and whether the differences in the GATES across all pairs of groups are zero. For the last test, we adjust p-values to account for multiple hypotheses testing using Holm's procedure and report the median of the adjusted p-values.

See also

Author

Riccardo Di Francesco